pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.2.4"
description = "Python page analyzer"
authors = ["Konstantin Chupin <chupin.k@yandex.ru>"]
packages = [
  {include = "page_analyzer"},
]
repository = "<https://github.com/Trankvill/python-project-83>"

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.2"
pytest = "^7.2.1"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
validators = "^0.20.0"
requests = "^2.28.2"
bs4 = "^0.0.1"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
